Moved widget tests to their own file
